fix: remove undocumented `probot.errorHandler` API
